go-runtime: allow empty staticdev package
When we build golang with dynamic linking (GO_DYNLINK=1),
the staticdev will be empty and so the package is not produced.
However the package is necessary to build the meta-go-toolchain sdk
so it is better to always produce it regardless of the value of GO_DYNLINK.

Fixes:
| *   - nothing provides go-runtime-staticdev needed by packagegroup-go-sdk-target-1.0-r0.all

This pacth is only need here in the meta-lts-mixins layer and is
not required upstream on oe-core/master.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>