Backport newer Python setuptools-rust and setuptools
To fix building python3-cryptography with Rust >= 1.83, a newer
python3-setuptools-rust is required.  As a consequence, newer
python3-setuptools is also required to satisfy its dependencies.
To simplify maintenance, the versions from scarthgap branch have
been used for the backport (1.9.0 and 69.1.1, respectively).

Note that extracting the specific cross-compilation fixes from
setuptools-rust was considered, but rejected due to the effort
required to get all of the required pieces untangled from its
commit history, and it not being obvious how to guarantee that
the result would not break some downstream usages.

Signed-off-by: Scott Murray <scott.murray@konsulko.com>