librsvg: upgrade 2.59.2 -> 2.60.0
0001-gdk-pixbuf-loader-meson.build-do-not-look-for-gdk-pi.patch and
0001-meson.build-do-not-force-disable-introspection-and-v.patch
patches were dropped, because they are included in this release.

======

Removed the CVE exception for CVE-2018-1000041, because it has been
fixed in 2017[1].

=====

During the do_configure step, the following error popped up:

| Program meson/query-rustc.py found: YES (/home/meeee/stuff/ptest-images/build/tmp/work/core2-64-poky-linux/librsvg/2.60.0/sources/librsvg-2.60.0/meson/query-rustc.py)
|
| ../sources/librsvg-2.60.0/meson.build:357:4: ERROR: Problem encountered: error[E0463]: can't find crate for `std`
|   |
|   = note: the `x86_64-poky-linux-gnu` target may not be installed
|   = help: consider downloading the target with `rustup target add x86_64-poky-linux-gnu`
|
| error: aborting due to 1 previous error
|
| For more information about this error, try `rustc --explain E0463`.

The meson script is trying to gather the static native libraries for the
target from rust, to link with them later. However it invokes rustc binary
directly without cargo, and the RUSTFLAGS environment variable is ignored,
which makes it to fail to find the Yocto targets.

To avoid this error, add a patch to append the RUSTFLAGS to the command
that calls rustc. (0001-query-rustc-append-RUSTFLAGS-to-rustc-executable.patch)

======

During the development of this version, using gdk-pixbuf-loader was enabled
for non-cross builds also. Unfortunately gdk-pixbuf-loader however has no
cross-wrapper script support officially, so it fails the build if it can't
find the executable (which is not present in the target build's sysroot).

To avoid to error, add a patch that reverts this change.
(0001-Revert-meson.build-do-not-force-disable-gdk-pixbuf-l.patch)

=====

Changelog:

- The minimum supported Rust version (MSRV) is 1.77.2.
- Many build fixes for various platforms (Amyspark, Alexander Kanavin,
  Colin Sane).
- #414 - Basic support for the dominant-baseline property (Paul Melis).
- #1146 - Parse the white-space property.  It is not processed yet;
  this is part of the work to support SVG2 text layout (Adetoye Anointing).
- #1138 - Report errors correctly from the parsers for attribute values.
- Portability fixes to the C header files.

Version 2.59.91
===============
- #1147 - Fix linkage on Illumos.
- #1143 - Fix the build of the static library on MinGW (Amyspark).
- Fix versioning of library name (Amyspark).
- Enable cross-compilation of the gdk-pixbuf loader if meson/qemu can
  handle it (Colin Sane).

Version 2.59.90
===============
- #414 - Basic support for the dominant-baseline property (Paul Melis).
- #1138 - Report errors correctly from the parsers for attribute values.
- #951 - Slightly improved test coverage (Vaishnavi Nawghare).
- #1136 - Don't use defined() in C macro definitions, since it is not
  portable.
- #1146 - Parse the white-space property.  It is not processed yet;
  this will come after the Outreachy internship (Adetoye Anointing).
- Documentation fixes (Kleis Auke Wolthuizen).
- Many build fixes (Amyspark).
- Improvements for cross builds (Alexander Kanavin).
- Fix CRLF characters in Python sources (sulincix).

[1]: https://gitlab.gnome.org/GNOME/librsvg/-/commit/c6ddf2ed4d768fd88adbea2b63f575cd523022ea
(this fixed the CVE according to https://nvd.nist.gov/vuln/detail/CVE-2018-1000041 )

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
(adapted from oe-core commit b0691bbbdf31b0e5c50ba319bdf6d16a278c8b5a)
Signed-off-by: Scott Murray <scott.murray@konsulko.com>