librsvg: add ptest support
The suite is quick, it takes ~5 seconds on my machine to execute.

The tests consist of 2 or 3 parts, depending on how one counts them.

The apitest PACKAGACONFIG compiles tests for the c-interface, as part of
the main meson build. This apitest expects test data to be present in
a folder relative the test binary: ../../rsvg/tests folder.

ptest-cargo compiles two sets of test: rsvg-convert tests, and generic assorted
tests for the library, in multiple binaries.

The rsvg-convert tests expect test data to be present in a "tests" folder,
in the same folder where the test binaries are.

The library tests expect the tests to be in a folder called "tests", that's
location is specified by the CARGO_MANIFEST_DIR environment variable.

As part of compiling the cargo tests, rsvg-convert is also compiled specifically
for the tests, however it is not supposed to be different from the actual binary.
It is deleted from the test set, because of two reasons:
1. It is not bit-identical to the actually installed rsvg-convert (it is compiled differently,
   with cargo only, without meson)
2. ptest-cargo class treated it as a test binary, and tried to execute it, which resulted
   in failing/hanging test.

It is avoided by not installing this binary, but rather use the real rsvg package
as a runtime dependency, and create a symlink to it in the run-ptest script.

Added liberation-fonts as a runtime dependency to the tests - the rsvg-convert
tests require some fonts to be available to be able to render during some
svg->pdf conversion tests.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
(adapted from oe-core commit 3aee16beac9b12403babfc7eef02be71b2297e70)
Signed-off-by: Scott Murray <scott.murray@konsulko.com>