go-helloworld: add from openembedded-core scarthgap
This allows a quick smoke test for whether go toolchain produces
working executables.

Last release where go-helloworld recipe was identical on scarthgap and
master is yocto-5.0.11. Afterwards license was fixed on scarthgap and
no further updates were done.
* fa45d6d5bec8fe503ff6b9166a3b4af31ea95369 go-helloworld: fix license
This commit is missing on master and will be submitted there.

Version copied here is from yocto-5.0.11 as a clean cherry-pick base.
* https://git.openembedded.org/openembedded-core/tree/meta/recipes-extended/go-examples?h=yocto-5.0.11

Signed-off-by: Peter Marko <peter.marko@siemens.com>